+++ c/dsls/Makefile.no-autoconfig
@@ -110,14 +110,20 @@ lemon: lemon.c \
   lemon_assert.c \
   lemon_error.h \
   lemon_error.c \
+  lemon_string.h \
+  lemon_string.c \
   lemon_memory.h \
   lemon_memory.c \
   lemon_option.h \
   lemon_option.c \
   lemon_structs.h \
   lemon_action.h \
-  lemon_action.c
-	$(DSLCC) -o lemon lemon.c lemon_assert.c lemon_error.c lemon_memory.c lemon_option.c lemon_action.c
+  lemon_action.c \
+  lemon_symbol.h \
+  lemon_symbol.c \
+  lemon_report.h \
+  lemon_report.c
+	$(DSLCC) -o lemon lemon.c lemon_assert.c lemon_error.c lemon_string.c lemon_memory.c lemon_option.c lemon_action.c lemon_symbol.c lemon_report.c
 
 # ----------------------------------------------------------------
 clean:
