+++ c/dsls/lemon_dims.h
@@ -1,0 +1,10 @@
+#ifndef LEMON_DIMS_H
+#define LEMON_DIMS_H
+
+#ifdef TEST
+#define MAXRHS 5       /* Set low to exercise exception code */
+#else
+#define MAXRHS 2000
+#endif
+
+#endif // LEMON_DIMS_H
